+++ src/database/src/generators/generator.rs
@@ -353,7 +353,7 @@ impl DatabaseGenerator {
                 },
                 board: ClubBoard::new(),
                 status: ClubStatus::Professional,
-                finance: ClubFinances::new(club.finance.balance, Vec::new()),
+                finance: ClubFinances::new(club.finance.balance as i64, Vec::new()),
                 academy: ClubAcademy::new(100),
                 colors: ClubColors {
                     background: club.colors.background.clone(),
